requirement.txt
<<<<<<< HEAD
astroid==3.0.1
asttokens==2.4.1
black==23.11.0
build==1.0.3
cachetools==5.3.2
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
check-manifest==0.49
click==8.1.7
colorama==0.4.6
coverage==7.3.2
dill==0.3.7
distlib==0.3.7
docutils==0.20.1
exceptiongroup==1.2.0
execnet==2.0.2
executing==2.0.1
filelock==3.13.1
greenlet==3.0.1
icecream==2.1.3
idna==3.6
importlib-metadata==6.9.0
iniconfig==2.0.0
isort==5.12.0
jaraco.classes==3.3.0
keyring==24.3.0
markdown-it-py==3.0.0
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.1.0
mypy-extensions==1.0.0
mysql-connector==2.2.9
-e git+ssh://git@github.com/ryanlong1004/python-starter-template.git@0067f795d058ac954c765c36e5b21b69c246a66a#egg=new_project
nh3==0.2.14
=======
black==23.11.0
click==8.1.7
greenlet==3.0.1
mypy-extensions==1.0.0
mysql-connector==2.2.9
>>>>>>> 850a7a039834cd7d5ec221180ab34a84ead8e3a9
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pathspec==0.11.2
<<<<<<< HEAD
pkginfo==1.9.6
platformdirs==4.0.0
pluggy==1.3.0
psycopg2-binary==2.9.9
Pygments==2.17.2
pylint==3.0.2
pyproject-api==1.6.1
pyproject_hooks==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
readme-renderer==42.0
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
six==1.16.0
SQLAlchemy==2.0.23
tomli==2.0.1
tomlkit==0.12.3
tox==4.11.4
twine==4.0.2
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
virtualenv==20.25.0
zipp==3.17.0
=======
platformdirs==4.0.0
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
SQLAlchemy==2.0.23
typing_extensions==4.8.0
tzdata==2023.3
>>>>>>> 850a7a039834cd7d5ec221180ab34a84ead8e3a9
